Code cleanup in pytorch models
