+++ forge/test/models/pytorch/vision/deit/test_deit.py
@@ -2,9 +2,7 @@
 
 # SPDX-License-Identifier: Apache-2.0
 import pytest
-import requests
 from datasets import load_dataset
-from PIL import Image
 from transformers import AutoFeatureExtractor, ViTForImageClassification
 
 import forge
@@ -22,8 +20,6 @@ def generate_model_deit_imgcls_hf_pytorch(variant):
     # STEP 3: Run inference on Tenstorrent device
     dataset = load_dataset("huggingface/cats-image")
     image_1 = dataset["test"]["image"][0]
-    url = "http://images.cocodataset.org/val2017/000000039769.jpg"
-    image_2 = Image.open(requests.get(url, stream=True).raw)
     img_tensor = image_processor(image_1, return_tensors="pt").pixel_values
     # output = model(img_tensor).logits
