+++ forge/test/models/pytorch/vision/vit/test_vit.py
@@ -2,9 +2,7 @@
 
 # SPDX-License-Identifier: Apache-2.0
 import pytest
-import requests
 from datasets import load_dataset
-from PIL import Image
 from transformers import AutoImageProcessor, ViTForImageClassification
 
 import forge
@@ -15,8 +13,6 @@ from test.utils import download_model
 
 dataset = load_dataset("huggingface/cats-image")
 image_1 = dataset["test"]["image"][0]
-url = "http://images.cocodataset.org/val2017/000000039769.jpg"
-image_2 = Image.open(requests.get(url, stream=True).raw)
 
 
 def generate_model_vit_imgcls_hf_pytorch(variant):
